pyproject.toml
[project]
name = "paperless-rules"
version = "0.1.0"
description = "Rule-based document classification and extraction for paperless-ngx"
readme = "PROJECT.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Yves" }]
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
dev = [
    "ruff>=0.5",
    "mypy>=1.8",
]

[project.scripts]
paperless-rules = "paperless_rules.cli:main"

[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paperless_rules"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q -m 'not e2e'"
asyncio_mode = "auto"
markers = [
    "e2e: tier-3 end-to-end tests requiring docker/podman compose + a paperless-ngx image (run with `pytest -m e2e`)",
]
